Extract shared workspace project navigation hook
