+++ src/client/routes/projects/workspaces/list.tsx
@@ -1,37 +1,13 @@
-import { useCallback } from 'react';
-import { useNavigate, useParams } from 'react-router';
 import { Loading } from '@/frontend/components/loading';
 import { trpc } from '@/frontend/lib/trpc';
 import { WorkspacesBoardView } from './components';
+import { useWorkspaceProjectNavigation } from './use-workspace-project-navigation';
 
 export default function WorkspacesListPage() {
-  const { slug = '' } = useParams<{ slug: string }>();
-  const navigate = useNavigate();
+  const { slug, projects, handleProjectChange, handleCurrentProjectSelect } =
+    useWorkspaceProjectNavigation();
 
   const { data: project } = trpc.project.getBySlug.useQuery({ slug });
-  const { data: projects } = trpc.project.list.useQuery({ isArchived: false });
-
-  const handleProjectChange = useCallback(
-    (value: string) => {
-      if (value === '__manage__') {
-        void navigate('/projects');
-        return;
-      }
-      if (value === '__create__') {
-        void navigate('/projects/new');
-        return;
-      }
-      void navigate(`/projects/${value}/workspaces`);
-    },
-    [navigate]
-  );
-
-  const handleCurrentProjectSelect = useCallback(() => {
-    if (!slug) {
-      return;
-    }
-    void navigate(`/projects/${slug}/workspaces`);
-  }, [navigate, slug]);
 
   if (!project) {
     return <Loading message="Loading project..." />;
